refactor: split telemetryvalidation controller by responsibility
